// src/common/enums/tipo_Transaccion.enum.ts
export enum Tipo_Transaccion {
    Guardar = "guardar",
    Guardar_Con_Parametros = "guardar_con_parametros",
    Actualizar = "actualizar",
    Actualizar_Con_Parametros = "actualizar_con_parametros",
    Eliminar = "eliminar",
    Eliminar_Con_Parametros = "eliminar_con_parametros",
    Consultar = "consultar",
    Consultar_Con_Parametros = "consultar_con_parametros"
}